Use subcommands (#196)
fixes to interviewer and suggestions
